requirements.txt
pandas>=1.5.0
jieba>=0.42.1
pyahocorasick>=1.4.4
openpyxl>=3.0.10
tqdm>=4.64.0
numpy>=1.21.0